wip: split vault crypto modules
